refactor: replace PNG images with WEBP format for improved performance and remove unused image files
