+++ app/(auth)/add-details/page.tsx
@@ -27,7 +27,7 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { addDetailsSchema } from '@/lib/auth-schema';
-import addDetails from '@/public/images/add-details.png';
+import addDetails from '@/public/images/add-details.webp';
 import AddDetailsMobile from '@/public/images/details-mobile-version.svg';
 import Image from 'next/image';
 import { redirect } from 'next/navigation';
